test: use new reverse proxy
